[not verified] refactor to shared
