+++ projects/plugins/jetpack/extensions/blocks/subscriptions/constants.js
@@ -1,5 +1,3 @@
-import { __ } from '@wordpress/i18n';
-
 export const MIN_BORDER_RADIUS_VALUE = 0;
 export const MAX_BORDER_RADIUS_VALUE = 50;
 export const DEFAULT_BORDER_RADIUS_VALUE = 0;
@@ -17,23 +15,3 @@ export const MAX_SPACING_VALUE = 50;
 export const DEFAULT_SPACING_VALUE = 10;
 
 export const DEFAULT_FONTSIZE_VALUE = '16px';
-
-export const META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS = '_jetpack_newsletter_access';
-
-export const accessOptions = {
-	everybody: {
-		key: 'everybody',
-		label: __( 'Everyone', 'jetpack' ),
-		panelHeading: __( 'Everyone', 'jetpack' ),
-	},
-	subscribers: {
-		key: 'subscribers',
-		label: __( 'Anyone subscribed', 'jetpack' ),
-		panelHeading: __( 'All subscribers', 'jetpack' ),
-	},
-	paid_subscribers: {
-		key: 'paid_subscribers',
-		label: __( 'Paid subscribers only', 'jetpack' ),
-		panelHeading: __( 'Paid subscribers', 'jetpack' ),
-	},
-};
