+++ projects/plugins/jetpack/extensions/blocks/subscriptions/edit.js
@@ -16,6 +16,7 @@ import { __, _n, sprintf } from '@wordpress/i18n';
 import classnames from 'classnames';
 import { isEqual } from 'lodash';
 import { getValidatedAttributes } from '../../shared/get-validated-attributes';
+import { isNewsletterFeatureEnabled } from '../../shared/memberships';
 import { getSubscriberCounts } from './api';
 import './view.scss';
 import defaultAttributes from './attributes';
@@ -29,7 +30,7 @@ import {
 import SubscriptionControls from './controls';
 import { SubscriptionsPlaceholder } from './subscription-placeholder';
 import SubscriptionSkeletonLoader from './subscription-skeleton-loader';
-import GetAddPaidPlanButton, { isNewsletterFeatureEnabled } from './utils';
+import GetAddPaidPlanButton from './utils';
 import { name } from './';
 
 const { getComputedStyle } = window;
