+++ projects/plugins/jetpack/extensions/blocks/subscriptions/settings.js
@@ -5,10 +5,12 @@ import { useSelect } from '@wordpress/data';
 import { PostVisibilityCheck, store as editorStore } from '@wordpress/editor';
 import { createInterpolateElement } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
+import {
+	META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS,
+	accessOptions,
+} from '../../../extensions/shared/memberships';
 import { store as membershipProductsStore } from '../../store/membership-products';
-import { META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS, accessOptions } from './constants';
 import { getPaidPlanLink, getShowMisconfigurationWarning, MisconfigurationWarning } from './utils';
-
 import './settings.scss';
 
 export function Link( { href, children } ) {
