+++ projects/plugins/jetpack/extensions/blocks/subscriptions/utils.js
@@ -1,8 +1,7 @@
-import { getJetpackData } from '@automattic/jetpack-shared-extension-utils';
 import { Button, ToolbarButton, Notice } from '@wordpress/components';
-import { useEntityProp } from '@wordpress/core-data';
 import { createInterpolateElement } from '@wordpress/element';
 import { _x, __ } from '@wordpress/i18n';
+import { accessOptions } from '../../shared/memberships';
 
 /**
  * Apply HTML encoding for special characters inside shortcode attributes.
@@ -27,10 +26,6 @@ export const getPaidPlanLink = alreadyHasNewsletterPlans => {
 	const link = 'https://wordpress.com/earn/payments-plans/' + location.hostname;
 	// We force the "Newsletters plan" link only if there is no plans already created
 	return alreadyHasNewsletterPlans ? link : link + '#add-newsletter-payment-plan';
-};
-
-export const isNewsletterFeatureEnabled = () => {
-	return getJetpackData()?.jetpack?.is_newsletter_feature_enabled ?? false;
 };
 
 export const getShowMisconfigurationWarning = ( postVisibility, accessLevel ) => {
@@ -74,16 +69,3 @@ export default function GetAddPaidPlanButton( { context = 'other', hasNewsletter
 		</Button>
 	);
 }
-
-export const GetAccessLevel = postType => {
-	const [ postMeta = [] ] = useEntityProp( 'postType', postType, 'meta' );
-
-	let accessLevel =
-		postMeta[ META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS ] ?? accessOptions.everybody.key;
-
-	// If accessLevel is ''
-	if ( ! accessLevel ) {
-		accessLevel = accessOptions.everybody.key;
-	}
-	return accessLevel;
-};
